Added Xbee Wifi frame parsing and building capability.
Refactored to enable the options.module feature. Which is required for different structured packets for the Wifi modules.

New analog and digital options for wifi modules, so now the MASK constant is per module.

Simple tests for the new wifi features.

Fixed the constructor to properly extend the default settings for options.

Needs test on an S6B module to see if the adsent application header is a firmware bug or just old documentation. CMD